Extract CreateFromImage.
