+++ src/features/linodes/LinodesCreate/LinodesCreate.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import Axios from 'axios';
-import * as moment from 'moment';
-import { flatten, tail, groupBy, pathOr } from 'ramda';
+import { pathOr } from 'ramda';
 
 import {
   withStyles,
@@ -12,28 +11,14 @@ import {
 import Typography from 'material-ui/Typography';
 import AppBar from 'material-ui/AppBar';
 import Tabs, { Tab } from 'material-ui/Tabs';
-import Grid from 'material-ui/Grid';
 
 import { API_ROOT } from 'src/constants';
-import TabbedPanel from 'src/components/TabbedPanel';
-import ExpandPanel from 'src/components/ExpandPanel';
-import SelectionCard from 'src/components/SelectionCard';
-import PromiseLoader, { PromiseLoaderResponse } from 'src/components/PromiseLoader';
 
-const distroIcons = {
-  Arch: 'archlinux',
-  CentOS: 'centos',
-  CoreOS: 'coreos',
-  Debian: 'debian',
-  Fedora: 'fedora',
-  Gentoo: 'gentoo',
-  openSUSE: 'opensuse',
-  Slackware: 'slackware',
-  Ubuntu: 'ubuntu',
-};
+import PromiseLoader, { PromiseLoaderResponse } from 'src/components/PromiseLoader';
+import CreateFromImage from './CreateFromImage';
 
 type Styles =
-'root';
+  'root';
 
 const styles = (theme: Theme & Linode.Theme): StyleRules => ({
   root: {
@@ -70,110 +55,22 @@ class LinodeCreate extends React.Component<FinalProps, State> {
   }
 
   handleImageClick = (event: React.MouseEvent<HTMLElement>, imageID: string) => {
+
     this.setState({ selectedImageID: imageID });
-  }
-
-  renderCreateFromImage = () => {
-    const images  = pathOr([], ['response', 'data'], this.props.images);
-
-    const publicImages = images.filter(
-      (image: Linode.Image) => image.id.startsWith('linode'));
-    const publicImagesByVendor = groupBy<Linode.Image>(
-      image => (image.vendor as string))(publicImages);
-    const firstImagesByVendor = Object.keys(publicImagesByVendor).map(vendor => (
-      (publicImagesByVendor[vendor] as Linode.Image[]).sort(
-        (imageA, imageB) => {
-          return (moment(imageB.created).diff(moment(imageA.created)));
-        })[0]));
-    const sortedFirstImagesByVendor = (firstImagesByVendor as Linode.Image[]).sort(
-      (imageA, imageB) => {
-        return Number(
-          (imageA.vendor as string).toLowerCase() > (imageB.vendor as string).toLowerCase());
-      });
-
-    const restImagesByVendor = flatten<Linode.Image>(
-      Object.keys(publicImagesByVendor).map(vendor => tail(
-        (publicImagesByVendor[vendor] as Linode.Image[]).sort(
-          (imageA, imageB) => {
-            return (moment(imageB.created).diff(moment(imageA.created)));
-          }))));
-    const sortedRestImagesByVendor = (restImagesByVendor as Linode.Image[]).sort(
-      (imageA, imageB) => {
-        return Number(
-          (imageA.vendor as string).toLowerCase() > (imageB.vendor as string).toLowerCase());
-      });
-
-    const privateImages = images.filter((image: Linode.Image) => image.id.startsWith('private'));
-
-    return (
-      <TabbedPanel
-        header="Select Image Type"
-        tabs={[
-          {
-            title: 'Public Images',
-            render: () => (
-              <React.Fragment>
-                <Grid container>
-                  {sortedFirstImagesByVendor.length
-                  && sortedFirstImagesByVendor.map((image: Linode.Image) => (
-                    <SelectionCard
-                      checked={image.id === this.state.selectedImageID}
-                      onClick={e => this.handleImageClick(e, image.id)}
-                      renderIcon={() => {
-                        return <span className={`fl-${distroIcons[(image.vendor as string)]}`} />;
-                      }}
-                      heading={(image.vendor as string)}
-                      subheadings={[image.label]}
-                    />
-                  ))}
-                </Grid>
-                <ExpandPanel name="Show Older Images">
-                  <Grid container>
-                    {sortedRestImagesByVendor.length
-                    && sortedRestImagesByVendor.map((image: Linode.Image) => (
-                      <SelectionCard
-                        checked={image.id === this.state.selectedImageID}
-                        onClick={e => this.handleImageClick(e, image.id)}
-                        renderIcon={() => {
-                          return <span className={`fl-${distroIcons[(image.vendor as string)]}`} />;
-                        }}
-                        heading={(image.vendor as string)}
-                        subheadings={[image.label]}
-                      />
-                    ))}
-                  </Grid>
-                </ExpandPanel>
-              </React.Fragment>
-            ),
-          },
-          {
-            title: 'My Images',
-            render: () => (
-              <Grid container>
-                {privateImages.length && privateImages.map((image: Linode.Image) => (
-                  <SelectionCard
-                    checked={image.id === this.state.selectedImageID}
-                    onClick={e => this.handleImageClick(e, image.id)}
-                    renderIcon={() => {
-                      return <span className="fl-tux" />;
-                    }}
-                    heading={(image.label as string)}
-                    subheadings={[(image.description as string)]}
-                  />
-                ))}
-              </Grid>
-            ),
-          },
-        ]}
-      >
-      </TabbedPanel>
-    );
   }
 
   tabs = [
     {
       title: 'Create from Image',
-      render: this.renderCreateFromImage,
+      render: () => {
+        const images  = pathOr([], ['response', 'data'], this.props.images);
+
+        return <CreateFromImage
+          images={images}
+          handleSelection={this.handleImageClick}
+          selectedImageID={this.state.selectedImageID}
+        />;
+      },
     },
   ];
